fix: UserScript pronounce 타입 변경

+++ be/ddobagi/src/main/java/com/a608/ddobagi/db/entity/UserScript.java
@@ -16,7 +16,7 @@ public class UserScript {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	private int pronounce;
+	private float pronounce;
 
 	@Column(name = "user_id")
 	private Long userId;
